requirements.txt
# BOLT INNOVATORS - Smart Waste Classification System
# Requirements

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Machine Learning
numpy==1.24.3

# Image Processing
Pillow==10.0.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0  # For production deployment

# External AI API Dependencies
google-generativeai>=0.3.0  # For Gemini API
openai>=1.0.0               # For ChatGPT API
requests>=2.28.0            # For API requests
asyncio>=3.4.3              # For asynchronous API calls